backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "The backend for SPARC"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==6.0.3",
    "django-cors-headers==4.9.0",
    "djangorestframework==3.17.0",
    "psycopg2-binary==2.9.11",
    "channels==4.3.2",
    "daphne==4.2.1",
    "pytest==9.0.2",
    "pytest-django==4.12.0",
    "pytest-cov==7.0.0",
    #"django-stubs[compatible-mypy]",
    #"mypy"
]

[tool.pyright]
typeCheckingMode = "standard"

[dependency-groups]
dev = [
    "django-types>=0.23.0",
]
